cheques basics improved...

+++ Turquaz/com/turquaz/cheque/ui/CheUIChequeCollectFromBankUpdate.java
@@ -18,7 +18,7 @@ package com.turquaz.cheque.ui;
 
 /**
  * @author Onsel
- * @version $Id: CheUIChequeCollectFromBankUpdate.java,v 1.3 2005/03/02 13:14:32 cemdayanik Exp $
+ * @version $Id: CheUIChequeCollectFromBankUpdate.java,v 1.4 2005/03/15 13:45:09 onsel Exp $
  */
 
 import java.util.ArrayList;
@@ -207,7 +207,7 @@ public class CheUIChequeCollectFromBankUpdate extends
 	            }   
 	           
 //	          TODO cheq trans exRate
-	           CheBLUpdateChequeRoll.updateChequeRollIn(chequeRoll,null,null,null,compChequeRoll.getTxtRollNo().getText().trim(),compChequeRoll.getDatePicker1().getDate(),chequeList,EngBLCommon.CHEQUE_TRANS_COLLECT_FROM_BANK,false,EngBLCommon.getBaseCurrencyExchangeRate());
+	           CheBLUpdateChequeRoll.updateChequeRollIn(chequeRoll,null,null,null,compChequeRoll.getTxtRollNo().getText().trim(),compChequeRoll.getDatePicker1().getDate(),chequeList,EngBLCommon.CHEQUE_TRANS_COLLECT_FROM_BANK.intValue(),false,EngBLCommon.getBaseCurrencyExchangeRate());
 	           EngUICommon.showMessageBox(getParent(),Messages.getString("CheUIChequeInPayroll.13"),SWT.ICON_INFORMATION); //$NON-NLS-1$
 	           isUpdated=true;
 	           dialogShell.close();
