+++ Turquaz/com/turquaz/cheque/ui/CheUIChequeInPayroll.java
@@ -17,7 +17,7 @@ package com.turquaz.cheque.ui;
 
 /**
 * @author  Onsel
-* @version  $Id: CheUIChequeInPayroll.java,v 1.17 2005/03/02 13:14:32 cemdayanik Exp $
+* @version  $Id: CheUIChequeInPayroll.java,v 1.18 2005/03/15 13:45:09 onsel Exp $
 */
 
 import java.util.ArrayList;
@@ -291,7 +291,7 @@ public class CheUIChequeInPayroll extends org.eclipse.swt.widgets.Composite impl
         CheBLSaveChequeTransaction.saveChequeRoll(accountPicker.getTurqAccountingAccount(),
         		(TurqCurrentCard)currentPicker.getData()
 				,null,txtRollNo.getText().trim(),datePicker1.getDate(),
-				chequeList,EngBLCommon.CHEQUE_TRANS_IN,btnSumTotals.getSelection(),
+				chequeList,EngBLCommon.CHEQUE_TRANS_IN.intValue(),btnSumTotals.getSelection(),
 				EngBLCommon.getBaseCurrencyExchangeRate());
         EngUICommon.showMessageBox(getShell(),Messages.getString("CheUIChequeInPayroll.13"),SWT.ICON_INFORMATION); //$NON-NLS-1$
         newForm();
