+++ Turquaz/com/turquaz/cheque/ui/CheUIChequeOutPayrollBankUpdate.java
@@ -17,7 +17,7 @@ package com.turquaz.cheque.ui;
 
 /**
 * @author  Onsel
-* @version  $Id: CheUIChequeOutPayrollBankUpdate.java,v 1.5 2005/03/02 19:30:49 onsel Exp $
+* @version  $Id: CheUIChequeOutPayrollBankUpdate.java,v 1.6 2005/03/15 13:45:09 onsel Exp $
 */
 
 import java.util.ArrayList;
@@ -260,7 +260,7 @@ public class CheUIChequeOutPayrollBankUpdate extends org.eclipse.swt.widgets.Dia
 					compChequeRoll.getTxtRollNo().getText().trim(),
 					compChequeRoll.getDatePicker1().getDate(),
 					chequeList,
-					EngBLCommon.CHEQUE_TRANS_OUT_BANK,
+					EngBLCommon.CHEQUE_TRANS_OUT_BANK.intValue(),
 					compChequeRoll.getBtnSumTotals().getSelection(),
 					EngBLCommon.getBaseCurrencyExchangeRate());
 	           EngUICommon.showMessageBox(getParent(),Messages.getString("CheUIChequeInPayroll.13"),SWT.ICON_INFORMATION); //$NON-NLS-1$
